requirements.txt
tensorflow-cpu>=2.13.0
tensorflow-hub>=0.14.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
roboflow>=1.1.12 